Removed Server APP and added networking
Modified alot of things related to networking and it should now be in a stable state to use ish. also added a editor project if i decide to make it a doom style multi player game instead.

+++ Networking/NetworkMessage.cs
@@ -8,7 +8,7 @@ namespace Networking
 {
     internal struct NetworkMessage
     {
-        public ushort PacketId;
+        public uint PacketId;
         public PacketFlag PacketFlags;
         public PacketType PacketType;
